+++ .gitignore
@@ -1,3 +1,4 @@
+cmake-build-debug
 build
 out
 .cache
