+++ headers/Finish.hpp
@@ -1,0 +1,12 @@
+#pragma once
+ 
+#include "SFML/Graphics.hpp"
+#include "CustomWall.hpp"
+
+/*
+* @brief A class that represents a finish area. Meant to be drawn to the screen.
+*/
+class Finish : public CustomWall {
+public:
+    Finish();
+};
