+++ headers/WallType.hpp
