+++ src/ClassicWall.cpp
@@ -1,0 +1,9 @@
+//
+// Created by luk27 on 05.03.2023.
+//
+#include "../headers/ClassicWall.hpp"
+
+ClassicWall::ClassicWall() {
+    type = WallType::ClassicWall;
+    setFillColor(sf::Color::Black);
+}
